move source code to sub directory
