+++ ppa_analysis/alt_ppa.py
