+++ ppa_analysis/battery.py
