+++ ppa_analysis/firming_contracts.py
