+++ ppa_analysis/get_load_and_weather_data.py
