+++ ppa_analysis/msat_replicator.py
@@ -1,16 +1,16 @@
 import pandas as pd
 import numpy as np
 import scenario_runner
-from getting_data.import_emissions_data import get_emissions_intensity
+from ppa_analysis.import_emissions_data import get_emissions_intensity
 
 
 # TODO: allow user input filenames and locations throughout this function: 
 # will need to consider what the user interface will look like / how users will
 # interact with the tool. Just through python, flask app...?
-load_profiles = pd.read_csv('tests/data/sample_one_day_data.csv')
-price_profiles = pd.read_csv('tests/data/sample_price_set.csv')
-charge_set = pd.read_csv('tests/data/sample_charge_set.csv')
-scenarios = pd.read_csv('tests/data/sample_scenarios.csv').fillna(np.nan).replace([np.nan], [None])
+load_profiles = pd.read_csv('../tests/data/sample_one_day_data.csv')
+price_profiles = pd.read_csv('../tests/data/sample_price_set.csv')
+charge_set = pd.read_csv('../tests/data/sample_charge_set.csv')
+scenarios = pd.read_csv('../tests/data/sample_scenarios.csv').fillna(np.nan).replace([np.nan], [None])
 emissions = get_emissions_intensity(
     start=load_profiles['DateTime'].iloc[0], 
     end=load_profiles['DateTime'].iloc[-1],
