+++ ppa_analysis/ppa.py
