+++ ppa_analysis/residuals.py
