+++ ppa_analysis/scenario_runner.py
@@ -1,7 +1,6 @@
 import pandas as pd
 import residuals
 import tariffs
-import getting_data.import_emissions_data as import_emissions_data
 import hybrid
 import ppa
 import numpy as np
